% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muddy_plotting.R
\name{hsla_to_rgba}
\alias{hsla_to_rgba}
\title{Convert HSLa values to RGBa values}
\usage{
hsla_to_rgba(h, s, l, a = 1, pct = F)
}
\arguments{
\item{h}{A value between 0 and 360 representing the whole degree of the hue}

\item{s}{A value between 0.0 - 1, (0 - 100\%) representing the saturation of the color}

\item{l}{A value between 0.0 - 1, (0 - 100\%) representing the lightness of the color}

\item{a}{A value between 0 and 1 representing the alpha of the color}

\item{pct}{An indicator of whether the rgb output values are scaled 0-255 or percentages. Defaults to \code{FALSE}. If rgb values are output as percentages, set the output will be a decimal between 0 and 1.}
}
\value{
A list of 4 values representing the RGBa values of the color
}
\description{
Convert HSLa values to RGBa values by using the HSLa color model.
}
