% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_aa_1}
\alias{media_aa_1}
\alias{media_aa_2}
\title{Calculate the Average Audience (AA)}
\usage{
media_aa_1(rating = NA, ue = NA, aa = NA)

media_aa_2(vpvh = NA, hh = NA, aa = NA)
}
\arguments{
\item{rating}{The percentage of a specific population group which is tuned to the Average Minute of a Program or Daypart.
\itemize{
\item One Rating Point = 1\% of the Population.
\item Five Rating Points = 5\% of the Population
\item Ten Rating Points = 10\% of the Population
}

Typically is expressed as a percentage (\%), but input as a decimal.}

\item{ue}{The total number of persons or homes in a population.
Measured in thousands (000s).}

\item{aa}{The average viewing for a program. Can be expressed as a projection (000s) or as a rating (\%).}

\item{vpvh}{A measure of a program’s audience composition (or profile) relative to the Households tuned to the program.}

\item{hh}{The projection of the household universal estimate. Measured in thousands (000s).}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
Average Audience reflects viewing for an average minute to a program and is an average of the audience at the specific minute (Min. 1, Min. 2, Min. 3, etc…).

Note: Impressions can be added together across demos, dayparts or stations/sources (ex: M18-49 + F18-49 = P18-49)
}
