% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_aa_rating_1}
\alias{media_aa_rating_1}
\alias{media_aa_rating_2}
\alias{media_aa_rating_3}
\title{Calculate the Average Audience (AA) Rating}
\usage{
media_aa_rating_1(share = NA, hut = NA, aa_rating = NA)

media_aa_rating_2(aa = NA, ue = NA, aa_rating = NA)

media_aa_rating_3(grps = NA, num_spots = NA, aa_rating = NA)
}
\arguments{
\item{share}{The Percent of Households Viewing Television that are tuned to a particular program/network during average minute of program or daypart. Typically is expressed as a percentage (\%), but input as a decimal.}

\item{hut}{The percentage of Total Television Households that are viewing Television during a given period. Typically expressed as a percentage (\%), but returned as a decimal.}

\item{aa_rating}{The percent of all TV Households or persons tuned to a specific station. Typically expressed as a percentage (\%), but returned as a decimal.}

\item{aa}{The projection of average viewing for a program. Measured in thousands (000s).}

\item{ue}{The total number of persons or homes in a population.
Measured in thousands (000s).}

\item{grps}{The sum of all ratings for all programs in an advertising schedule. One rating point equals one percent of total audience (universe).}

\item{num_spots}{The number of advertising/program spots.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The estimate size of the Television audience relative to the total universe, expressed as a percentage. The percent of all TV Households or persons tuned to a specific station.

Ratings are NOT additive across different Demos; Ratings must be weight-average when combined.
}
