% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{media_abbreviations}
\alias{media_abbreviations}
\title{Media Abbreviations}
\format{
A data frame with 11 rows and 2 variables:
\describe{
\item{abbreviation}{A commonly used abbreviation in the media world.}
\item{meaning}{The longform explanation of what the short name represents.}
}
}
\source{
Nielsen Company.
}
\usage{
media_abbreviations
}
\description{
This dataset gives translations between common media acronyms and abbreviations, to what they represent.
}
\keyword{datasets}
