% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_avg_hours}
\alias{media_avg_hours}
\title{Calculate Average Hours of Viewing}
\usage{
media_avg_hours(duration = NA, hut_put = NA, avg_hours = NA)
}
\arguments{
\item{duration}{Duration of period.}

\item{hut_put}{The percentage of Total Persons / Households in a particular demographic group that are viewing Television during a given time period.}

\item{avg_hours}{HUT / PUT converted to the average hours of viewing per home or per person.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
HUT / PUT converted to the average hours of viewing per home or per person.
}
