% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_coverage}
\alias{media_coverage}
\title{Calculate the Coverage}
\usage{
media_coverage(cvg_projection = NA, cvg_ue = NA, cvg_rating = NA)
}
\arguments{
\item{cvg_projection}{The projected coverage, measured in thousands (000s).}

\item{cvg_ue}{The universal estimate of coverage, measured in thousands (000s).}

\item{cvg_rating}{The percentage of homes or persons able to receive an individual network or channel. Typically expressed as a percentage (\%), but returned as a decimal.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The percentage of homes or persons able to receive an individual network or channel.
}
