% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_cpm}
\alias{media_cpm}
\title{Calculate Cost per Thousand (CPM)}
\usage{
media_cpm(cost = NA, gross_aa = NA, cpm = NA)
}
\arguments{
\item{cost}{Media cost (in dollars).}

\item{gross_aa}{Gross audience (AA) who watched the program or advertisement. Measured in thousands (000s).}

\item{cpm}{The cost of delivering One Thousand Impressions within a defined population group.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The cost of delivering One Thousand Impressions within a defined population group.

Note:
\itemize{
\item Basic Formula of negotiation among cable networks
\item Measures efficiencies of media schedule
\item Allows for cross media evaluation
}

In other words, the CPM can be boiled down to a tongue-in-cheek formula:
CPM = \eqn{\frac{Bucks}{Schmucks}}
}
