% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_cpp_1}
\alias{media_cpp_1}
\alias{media_cpp_2}
\title{Cost Per Point (CPP)}
\usage{
media_cpp_1(u_cost = NA, rating = NA, cpp = NA)

media_cpp_2(s_cost = NA, grps = NA, cpp = NA)
}
\arguments{
\item{u_cost}{Average unit cost (in dollars).}

\item{rating}{The percentage of a specific population group which is tuned to the Average Minute of a Program or Daypart.
\itemize{
\item One Rating Point = 1\% of the Population.
\item Five Rating Points = 5\% of the Population
\item Ten Rating Points = 10\% of the Population
}

Typically is expressed as a percentage (\%), but input as a decimal.}

\item{cpp}{The cost to deliver a single rating point (1\% of the defined population).}

\item{s_cost}{Total Schedule Cost (in dollars).}

\item{grps}{The sum of all ratings for all programs in an advertising schedule. One rating point equals one percent of total audience (universe).}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The cost to deliver a single rating point (1\% of the defined population).
}
