% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{media_demo_conversions}
\alias{media_demo_conversions}
\title{Media Demo Conversions}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{target_demo}{A sex-agnostic demo age range.}
\item{formula}{A string describing the operations needed to calculate the given demo by other demos.}
}
}
\source{
Nielsen Company.
}
\usage{
media_demo_conversions
}
\description{
This dataset gives instructions on how to convert between certain media demo blocks.
}
\keyword{datasets}
