% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_frequency}
\alias{media_frequency}
\title{Calculate Frequency}
\usage{
media_frequency(grps = NA, frequency = NA, reach = NA)
}
\arguments{
\item{grps}{The sum of all ratings for all programs in an advertising schedule. One rating point equals one percent of total audience (universe).}

\item{frequency}{The average number of times that each Household (or Person) is exposed to an advertising schedule or campaign.}

\item{reach}{The number of different households or persons who are exposed at least once to a program, daypart, or advertising schedule over a given period of time.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The average number of times that each Household (or Person) is exposed to an advertising schedule or campaign. (Expressed as an absolute number).
}
