% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_gross_imp}
\alias{media_gross_imp}
\title{Calculate Gross Impressions}
\usage{
media_gross_imp(gross_imp = NA, ue = NA, grps = NA)
}
\arguments{
\item{gross_imp}{The total number of Households, or Persons, exposed to an advertising schedule. Measured in thousands (000s).}

\item{ue}{The total number of persons or homes in a population.
Measured in thousands (000s).}

\item{grps}{The sum of all ratings for all programs in an advertising schedule. One rating point equals one percent of total audience (universe).}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The total number of Households, or Persons, exposed to an advertising schedule.
}
