% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_grp_1}
\alias{media_grp_1}
\alias{media_grp_2}
\alias{media_grp_3}
\title{Calculate Gross Rating Points (GRPs)}
\usage{
media_grp_1(impressions = NA, ue = NA, grps = NA)

media_grp_2(rating = NA, num_spots = NA, grps = NA)

media_grp_3(reach = NA, frequency = NA, grps = NA)
}
\arguments{
\item{impressions}{The total number of Households, or Persons, exposed to an advertising schedule. Measured in thousands (000s).}

\item{ue}{The total number of persons or homes in a population.
Measured in thousands (000s).}

\item{grps}{The sum of all ratings for all programs in an advertising schedule. One rating point equals one percent of total audience (universe).}

\item{rating}{The percentage of a specific population group which is tuned to the Average Minute of a Program or Daypart.
\itemize{
\item One Rating Point = 1\% of the Population.
\item Five Rating Points = 5\% of the Population
\item Ten Rating Points = 10\% of the Population
}

Typically is expressed as a percentage (\%), but input as a decimal.}

\item{num_spots}{The number of advertising/program spots.}

\item{reach}{The number of different households or persons who are exposed at least once to a program, daypart, or advertising schedule over a given period of time; Also referred to as Cume, Unduplicated Audience or Net Audience.}

\item{frequency}{The average number of times that each Household (or Person) is exposed to an advertising schedule or campaign. (Expressed as an absolute number).}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The sum of all ratings for all programs in an advertising schedule. One rating point equals one percent of total audience (universe).

Note:
\itemize{
\item The sum of all rating points in a given schedule
\item Takes into account duplication
\item Can exceed 100
\item Describes the amount of media weight
}
}
