% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_hut_1}
\alias{media_hut_1}
\alias{media_hut_2}
\title{Calculate Households Using Television (HUT)}
\usage{
media_hut_1(hh_rating = NA, share = NA, hut = NA)

media_hut_2(hh_in_use = NA, total_hh_ue = NA, hut = NA)
}
\arguments{
\item{hh_rating}{The percentage of a households which are tuned to the Average Minute of a Program or Daypart.
\itemize{
\item One Rating Point = 1\% of the Population.
\item Five Rating Points = 5\% of the Population
\item Ten Rating Points = 10\% of the Population
}

Typically is expressed as a percentage (\%), but input as a decimal.}

\item{share}{The percentage of Households Viewing Television that are tuned to a particular
program/network during average minute of program or daypart. Typically expressed as a percentage (\%), but returned as a decimal.}

\item{hut}{The percentage of Total Television Households that are viewing Television during a given period. Typically expressed as a percentage (\%), but returned as a decimal.}

\item{hh_in_use}{The number of households with their TV sets in use.}

\item{total_hh_ue}{The total Universal Estimate for households.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The percentage of Total Television Households that are viewing Television during a given period.
}
