% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_put_1}
\alias{media_put_1}
\alias{media_put_2}
\title{Calculate Persons Using Television (PUT)}
\usage{
media_put_1(persons_viewing = NA, total_person_ue = NA, put = NA)

media_put_2(demo_rating = NA, demo_share = NA, put = NA)
}
\arguments{
\item{persons_viewing}{The number of persons viewing TV.}

\item{total_person_ue}{The total Universal Estimate of persons who watch TV.}

\item{put}{The percentage of Total Persons in a particular demographic group that are viewing Television during a given time period. Typically expressed as a percentage (\%), but returned as a decimal.}

\item{demo_rating}{The rating percentage of a program within a certain demographic. Typically expressed as a percentage (\%), but returned as a decimal.}

\item{demo_share}{The share percentage of a program within a specific demographic. Typically expressed as a percentage (\%), but returned as a decimal.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The percentage of Total Persons in a particular demographic group that are viewing Television during a given time period.
}
