% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_share}
\alias{media_share}
\title{Calculate the Share}
\usage{
media_share(rating = NA, hut = NA, share = NA)
}
\arguments{
\item{rating}{The percentage of a specific population group which is tuned to the Average Minute of a Program or Daypart.
\itemize{
\item One Rating Point = 1\% of the Population.
\item Five Rating Points = 5\% of the Population
\item Ten Rating Points = 10\% of the Population
}

Typically is expressed as a percentage (\%), but input as a decimal.}

\item{hut}{The percentage of Total Television Households that are viewing Television during a given period. Typically expressed as a percentage (\%), but returned as a decimal.}

\item{share}{The Percent of Households Viewing Television that are tuned to a particular program/network during average minute of program or daypart. Typically is expressed as a percentage (\%), but input as a decimal.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The Percent of Households Viewing Television that are tuned to a particular program/network during average minute of program or daypart
}
