% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_ue}
\alias{media_ue}
\title{Calculate the Universal Estimate}
\usage{
media_ue(impressions = NA, rating = NA, ue = NA)
}
\arguments{
\item{impressions}{The total number of households or persons exposed to an advertising schedule.
Measured in thousands (000s).}

\item{rating}{The percentage of a specific population group which is tuned to the Average Minute of a Program or Daypart.
\itemize{
\item One Rating Point = 1\% of the Population.
\item Five Rating Points = 5\% of the Population
\item Ten Rating Points = 10\% of the Population
}

Typically is expressed as a percentage (\%), but input as a decimal.}

\item{ue}{The total number of persons or homes in a population.
Measured in thousands (000s).}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
The total number of persons or homes in a population.
}
