% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_math.R
\name{media_vpvh}
\alias{media_vpvh}
\title{Calculate Viewers per Viewing Household (VPVHs)}
\usage{
media_vpvh(persons = NA, hh = NA, vpvh = NA)
}
\arguments{
\item{persons}{The projected number of persons viewing a program. Expressed in the thousands (000s).}

\item{hh}{The projected number of households viewing a program. Expressed in the thousands (000s).}

\item{vpvh}{A measure of a program’s audience composition (or profile) relative to the Households tuned to the program.}
}
\value{
The parameter whose input was left as \code{NA}.
}
\description{
A measure of a program’s audience composition (or profile) relative to the Households tuned to the program.

Note:
\itemize{
\item Shows the audience skew of a program/network
\item Can be expressed in Hundreds (.10) or Thousands (.100)
\item Is additive between demographics
\item Used in estimating a program’s target audience
}
}
