% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muddy_plotting.R
\name{rgba_to_hsla}
\alias{rgba_to_hsla}
\title{Convert RGBa values to HSLa values}
\usage{
rgba_to_hsla(r, g, b, a = 1, pct = F)
}
\arguments{
\item{r}{An integer between 0 and 255 representing the red value of the color}

\item{g}{An integer between 0 and 255 representing the green value of the color}

\item{b}{An integer between 0 and 255 representing the blue value of the color}

\item{a}{A value between 0 and 1 representing the alpha of the color}

\item{pct}{An indicator of whether the rgb values are percentages or scaled 0-255. Defaults to \code{FALSE}. If rgb values are input as percentages, set the parameters equal to a decimal between 0 and 1.}
}
\value{
Returns h/s/l/a in a vector, h = 0-360 deg, s = 0.0 - 1 (0-100\%), l = 0.0 - 1 (0-100\%)
}
\description{
Convert RGBa values to HSLa values by using the RGBa color model.
}
