% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{streak_counter}
\alias{streak_counter}
\title{Calculate Streaks}
\usage{
streak_counter(df, bool_column, date_column)
}
\arguments{
\item{df}{A dataframe object which has at least two columns.}

\item{bool_column}{A column of df which has a numeric boolean indication (0s and 1s)
of the streak to detect.}

\item{date_column}{A column of df which is used for ordering (be it a date, number, etc.).}
}
\value{
A copy of the dataframe object with a streak column added.
}
\description{
This function takes a dataframe, which has an ordering column and a numeric
boolean indicator (0s and 1s), and calculates the streak of the indicator.
This allows for streak counting and analysis, which can help determine
relationships over time.
}
